google-compute package is distinct for xenial
- the packages that are used for trusty are actually deprecated
xenial now installs the packages in teh way recommended by
google at this link

https://github.com/GoogleCloudPlatform/compute-image-packages#daemons

[#152702490](https://www.pivotaltracker.com/story/show/152702490)

+++ stemcell_builder/stages/system_google_packages/apply.sh
@@ -16,15 +16,31 @@ mkdir -p $chroot/tmp/google
 os_type="$(get_os_type)"
 if [ "${os_type}" == "ubuntu" ]
 then
-  # Copy google daemon packages into chroot
-  cp -R $assets_dir/google-ubuntu/*.deb $chroot/tmp/google/
+  if [ "${DISTRIB_CODENAME}" == "trusty" ]; then
+    # Copy google daemon packages into chroot
+    cp -R $assets_dir/google-ubuntu/*.deb $chroot/tmp/google/
 
-  run_in_chroot $chroot "apt-get update"
-  run_in_chroot $chroot "apt-get install -y python-setuptools python-boto"
-  run_in_chroot $chroot "dpkg --unpack /tmp/google/*.deb"
-  run_in_chroot $chroot "rm /var/lib/dpkg/info/google-compute-engine-init-trusty.postinst"
-  run_in_chroot $chroot "dpkg --configure google-compute-engine-init-trusty google-config-trusty google-compute-engine-trusty"
-  run_in_chroot $chroot "apt-get install -yf"
+    run_in_chroot $chroot "apt-get update"
+    run_in_chroot $chroot "apt-get install -y python-setuptools python-boto"
+    run_in_chroot $chroot "dpkg --unpack /tmp/google/*.deb"
+    run_in_chroot $chroot "rm /var/lib/dpkg/info/google-compute-engine-init-trusty.postinst"
+    run_in_chroot $chroot "dpkg --configure google-compute-engine-init-trusty google-config-trusty google-compute-engine-trusty"
+    run_in_chroot $chroot "apt-get install -yf"
+  elif [ "${DISTRIB_CODENAME}" == "xenial" ]; then
+    run_in_chroot $chroot "curl https://packages.cloud.google.com/apt/doc/apt-key.gpg | apt-key add -"
+
+    run_in_chroot $chroot "tee /etc/apt/sources.list.d/google-cloud.list << EOM
+deb http://packages.cloud.google.com/apt google-compute-engine-stretch-stable main
+deb http://packages.cloud.google.com/apt google-cloud-packages-archive-keyring-stretch main
+EOM"
+
+    pkg_mgr install "google-cloud-packages-archive-keyring"
+    pkg_mgr install "python-google-compute-engine=2.7.1-1"
+    pkg_mgr install "python3-google-compute-engine=2.7.1-1"
+    pkg_mgr install "google-compute-engine"
+
+    run_in_chroot $chroot "sed -i 's/metadata.google.internal/169.254.169.254/g' /usr/lib/python3/dist-packages/google_compute_engine/metadata_watcher.py"
+  fi
 
   # Hack: replace google metadata hostname with ip address (bosh agent might set a dns that it's unable to resolve the hostname)
   run_in_chroot $chroot "sed -i 's/metadata.google.internal/169.254.169.254/g' /usr/lib/python2.7/dist-packages/google_compute_engine/metadata_watcher.py"
